+++ test/sandbox/app/components/url_for_mailer_component.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class UrlForMailerComponent < ViewComponent::Base
+end