+++ test/test_helper.rb
@@ -95,8 +95,8 @@ def with_custom_component_path(new_value, &block)
   with_generate_option(:path, new_value, &block)
 end
 
-def with_custom_component_parent_class(new_value, &block)
-  with_generate_option(:component_parent_class, new_value, &block)
+def with_custom_parent_class(new_value, &block)
+  with_generate_option(:parent_class, new_value, &block)
 end
 
 def with_application_component_class