+++ docs/CHANGELOG.md
@@ -10,6 +10,10 @@ nav_order: 6
 
 ## main
 
+* BREAKING: Remove deprecated `use_helper(s)`. Use `include MyHelper` or `helpers.` proxy instead.
+
+    *Joel Hawksley*
+
 ## 4.0.0.alpha6
 
 * BREAKING: Remove `config.test_controller` in favor of `vc_test_controller_class` test helper method.