+++ docs/CHANGELOG.md
@@ -14,6 +14,10 @@ nav_order: 6
 
     *Joel Hawksley*
 
+* Validate compatibility with `Dry::Initializer` in test suite.
+
+    *Joel Hawksley*
+
 ## 4.0.0.alpha6
 
 * BREAKING: Remove `config.test_controller` in favor of `vc_test_controller_class` test helper method.