+++ docs/CHANGELOG.md
@@ -18,6 +18,10 @@ nav_order: 6
 
     *Joel Hawksley*
 
+* BREAKING: Rename `config.generate.component_parent_class` to `config.generate.parent_class`.
+
+    *Joel Hawksley*
+
 * `config.previews.default_layout` should default to nil.
 
     *Joel Hawksley*
@@ -36,7 +40,7 @@ nav_order: 6
 
     *Joel Hawksley*
 
-* BREAKING: `config.component_parent_class` is now `config.generate.parent_class`, moving the generator-specific option to the generator configuration namespace.
+* BREAKING: `config.component_parent_class` is now `config.generate.component_parent_class`, moving the generator-specific option to the generator configuration namespace.
 
     *Joel Hawksley*