+++ docs/CHANGELOG.md
@@ -10,6 +10,8 @@ nav_order: 6
 
 ## main
 
+## 4.0.0.alpha7
+
 * BREAKING: Remove deprecated `use_helper(s)`. Use `include MyHelper` or `helpers.` proxy instead.
 
     *Joel Hawksley*